Split out table ui component
